+++ src/main/kotlin/no/nav/cv/eures/konverterer/EducationHistoryConverter.kt
@@ -3,10 +3,7 @@ package no.nav.cv.eures.konverterer
 import no.nav.arbeid.cv.avro.Utdannelse
 import no.nav.cv.eures.model.*
 import no.nav.cv.eures.samtykke.Samtykke
-import java.text.Normalizer
-import java.text.SimpleDateFormat
-import java.time.LocalDate
-import java.util.*
+
 
 
 class EducationHistoryConverter(
@@ -30,14 +27,4 @@ class EducationHistoryConverter(
                     it.tilTidspunkt.toFormattedDateTime()),
             educationLevelCode = EducationLevelCode(code = it.nuskodeGrad.substring(0, 1))
     ) }
-
-    private fun LocalDate.toFormattedDateTime()
-            = FormattedDateTime(toString())
-
-    private fun Long.toFormattedDateTime() : FormattedDateTime {
-        val jdf = SimpleDateFormat("yyyy-MM-dd")
-        val date = Date(this)
-
-        return FormattedDateTime(jdf.format(date))
-    }
 }
