+++ src/main/kotlin/no/nav/cv/eures/model/Employment.kt
@@ -1,29 +1,17 @@
 package no.nav.cv.eures.model
 
-import no.nav.arbeid.cv.avro.Cv
-import no.nav.cv.eures.samtykke.Samtykke
-
-class Employment (
-        private val cv: Cv,
-        private val samtykke: Samtykke
-){
-    fun getEmploymentHistory() : EmploymentHistory {
-        if(!samtykke.utdanning) // TODO Ikke utdanning her
-            return EmploymentHistory(listOf())
-
-        return EmploymentHistory(listOf())
-    }
-}
+import com.fasterxml.jackson.dataformat.xml.annotation.JacksonXmlElementWrapper
 
 data class EmploymentHistory(
+        @JacksonXmlElementWrapper(useWrapping = false)
         val employerHistory: List<EmployerHistory>
 )
 
 data class EmployerHistory(
     val organizationName: String,
-    val organizationContact: PersonContact, // TODO Usikker paa denne mappingen
-    val industryCode: IndustryCode,
-    val employmentPeriod: AttendancePeriod
+    //val organizationContact: PersonContact, // TODO Usikker paa denne mappingen
+    val employmentPeriod: AttendancePeriod,
+    val industryCode: IndustryCode // TODO Mapping Styke&Janzz
 )
 
 data class IndustryCode(val value: String)
